i#7270 Ubuntu22: sample.callstack output mismatch fix (#7336)
The `code_api|sample.callstack` test fails to match the expected output
on both x86_64 and aarch64 with Ubuntu 22.04.
This test is supposed to use `libcallstack` to track `malloc()` calls, but
the binaries we run the test on, which come from eflags.c and nzcv.c,
don't call `malloc()` directly.
Depending on the GLIBC version, it looks like `INIT()` in eflags.c and
nzcv.c might call `malloc()` indirectly, but even when it does, `drwrap`
(used in `libcallstack`) is unable to find it.
We filed a separate issue #7359 on this.
To fix this test we build a new binary from: suite/tests/common/alloc.c,
which calls `malloc()` directly.

Issue: #7270 #7359