i#7360: Fix ldscript creation with binutils 2.44 and preferred base set
We used to adjust current location pointer to our preferred base address
at start of the linkscript, which creates a hole in the address space
and allows the linker to put the ELF header and other sections before
address of __executable_start symbol. This causes assertion failures
when building with binutils 2.44,

	Internal Error: DynamoRIO debug check failure:
	/dynamorio/core/unix/os.c:9618 check_start == dynamo_dll_start

Since binutils ld.bfd always determines the base address by

	. = SEGMENT_START("text-segment", 0x400000) + SIZEOF_HEADERS;

starting from at least 2003, it should be fine to drop the extra
adjustment and only replace the default base with our preferred base.

Fixes: #7360