i#7270 Ubuntu22: Migrate x86-{32,64}, android-arm-cross, a64-on-x86 (#7370)
Removes the restriction on the test set run on Ubuntu22, and removes the
UBUNTU_22 set itself.

Removes the x86-64 and x86-32 Ubuntu 20 Github Actions CI workflows,
since we have others than run the same configuration on Ubuntu 22.
Migrates the android-arm-cross-compile and a64-on-x86 to Ubuntu 22.

Removes the -ubuntu22 suffix from the workflows where that's the only
variation left now.

This has been made possible after a bunch of work on #7270 by other
contributors: @derekbruening, @edeiana, and @ivankyluk.

Issue: #7270