i#7376: Fix schedule_stats crash at verbose 4
Fixes a crash in the drmemtrace schedule_stats tool at verbose 4 where
it tries to get the input stream on an idle sequence.

Tested on the new threadsig checked-in trace.  Unfortunately the unit
tests mock the stream interfaces and so can't reproduce this, and the
top-level at verbose 4 with a big enough trace to hit idle is too slow
for a good regression test.

Fixes #7376