i#7381: Record syscall failure stats in syscall_mix tool
Augments the syscall_mix tool to record a map from syscall failure
codes to counts. The caveat is that the trace only contains this
information explicitly for syscall that were traced with
`-record_syscall`.

Generalizes the syscall_mix results in a statistics_t struct with an accessor.

Adds testing to the burst_syscall_inject test by adding failing
syscall, running syscall_mix, accessing its stats, and confirming the
results.

Fixes #7381