i#7355 per-sys-stats: Add per-syscall-number whether-switch latency histograms (#7391)
Augments the schedule_stats drmemtrace tool to keep two histograms for
each system call number: one for latencies of system calls causing a
context switch, and one for latencies not switching. Unfortunately the
latencies are all 0 for core-sharded-on-disk traces; we live with that
for now.

Generalizes the histogram support through virtual methods
create_histogram() and find_or_add_histogram() to make it easier to use
different histograms in subclasses with all the new histograms being
added here.

Adds a unit test and augments top-level test expected output.

Fixes #7355