i#7401: Ensure unique pid,tid in drmemtrace schedules
In dynamic drmemtrace schedules combining multiple workloads, there
can be pid or tid collisions.  We solve that by having the scheduler
set the workload ordinal in the top 32 bits of both the pid and tid.
We ignore tids being 64-bit values on Mac; and we do not support
unique values for 32-bit builds when trace_entry_t-based tools such as
record_filter are used.

Documents this, along with missing docs for cpuid and timestamp
modifications.

Updates the view tool and schedule stats to print tids in hex, to make
it easier to distinguish the workload ordinal.  Updates test templates
and docs for this change.

Adds scheduler unit tests.

Here is what this looks like:
===========================================================================
$ bin64/drrun -t drmemtrace -tool view -multi_indir ../src/clients/drcachesim/tests/drmemtrace.threadsig.x64.tracedir:../src/clients/drcachesim/tests/drmemtrace.threadsig.x64.tracedir -core_sharded -cores 1 -sched_quantum 10 2>&1 | less
Output format:
<--record#-> <--instr#->: <---tid---> <record details>
------------------------------------------------------------
           1           0:       d51be <marker: version 7>
           2           0:       d51be <marker: filetype 0xe40>
           3           0:       d51be <marker: cache line size 64>
...
------------------------------------------------------------
         135          60:   1000d51be <marker: version 7>
         136          60:   1000d51be <marker: filetype 0xe40>
         137          60:   1000d51be <marker: cache line size 64>
===========================================================================

Fixes #7401