i#6662 public traces: skip info.textproto and update doc.
Now that we added `info.textproto` files to public traces,
we need to skip them when iterating over the content of a
trace directory to avoid confusing the scheduler, as these
files are meant for human-consumption only, not tools.

We update the "Google workload traces" web page
describing the intended use of `info.textproto`.
We also clarify that memory operands in #DR_ISA_REGDEPS
instructions can only be inferred from their subsequent
read/write records.

Issue #6662