Update elfutils version to 0.191 release
We currently pin the elfutils project to a floating ref. To make DR
better packagable, we pin the elfutils commit to the next official
release that contains the current commit (which is 0.191). An update to
the latest release (0.192) has been considered, but requires rework of
the build infrastructure and in-depths testing.

The 0.191 release also applied the elf_memory.c fix, so we drop it here.
Note, that this is just preliminary as this patch needs to be
reintroduced when updating to elfutils-0.192.

  cc44ac67 libelf: Treat elf_memory as if using ELF_C_READ_MMAP

The other changes between c1058da5 and elfutils-0.191 seem not be be
affecting DR.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>