i#6662 public traces: Fix regdeps crash and assert (#7422)
Fixes a backward assert in instr_set_operation_size().

Fixes a crash if a regdeps instruction is disassembled with Intel, AT&T,
or Arm styles selected.

Adds a test of both issues.

Also tested on drdisas temporarily disabling the "return 1" with a
command that hits a SIGSEGV without this fix; with the fix it just
doesn't print operands (the SYSLOG in this PR is only visible in full-DR
runs):
```
$ clients/bin64/drdisas -syntax intel -mode regdeps 00001931 04020204 00000026
'regdeps' mode does not support 'intel' syntax
 00001931 04020204 load store [4byte]
 00000026
```

Issue: #6662