i#6495 syscall inject: Fix retaddr tracking inside kernel trace (#7424)
Fixes the tracking for return addresses in drmemtrace invariant_checker
for system call and context switch traces.

This fixes "Function marker retaddr should match prior call" in traces
injected with kernel syscall sequences. This error weirdly manifested in
the user-space part of the trace, but only after the trace was injected
with kernel sequences. See details below.

At call instructions, the invariant_checker pushes the fallthrough pc to
the retaddr stack, which is compared with the value of the later
function return address marker. This entry is popped from the retaddr
stack when we see a return instruction. At kernel_event markers, the
invariant_checker inserts a dummy zero entry in the retaddr stack which
is expected to be popped when the signal handler's return instruction is
seen. This zero entry is not used for any check, but is used anyway
simply so that we can treat all returns the same way.

We have some existing logic that skips pushing and popping from the
retaddr stack for the kernel parts of the trace (since we don't do
function tracing in the kernel, we don't need to track retaddrs for the
kernel trace). But there was logic missing to do the same for the dummy
zero entry at kernel_event markers inside the kernel trace.

Adds a unit test (derived from a real trace) which fails without this
fix. It includes a tail call, which does not push anything to the
retaddr stack as it is a non-call instr. Without this fix, the top of
stack held the zero entry from the prior syscall trace, which caused the
function return address marker check to fail. With this fix, the top of
stack holds the return address of the prior call-instr (from before the
syscall trace), which is what the tail call also returns to, so the check
passes.

Issue: #6495, #7157