i#7429: Allow "aux" directory inside the trace directory
To avoid bumping DynamoRIO's version whenever we add a new
file to the trace directory (due to the scheduler reporting
an error when encountering an unrecognized file), we introduce
a new `aux` directory that is ignored by the scheduler when
going through the trace files inside the trace directory.
So, `aux` can contain auxiliary trace files (e.g., the
human-readable `info.textproto` file) that the scheduler
does not need to explicitely know about.

We add `aux/info.textproto` to
`clients/drcachesim/tests/drmemtrace.threadsig.x64.tracedir/aux/info.textproto`
(one of the test trace directories) which needs to be ignored
when running end-to-end tests that use this trace (e.g.,
tool.record_filter_as_traced).

Issue #7429