i#2154 Android64: Support building on macOS hosts (#7431)
This makes two small changes to the CMake files to enable building for
Android on non-Linux host machines:

- toolchain-android-llvm.cmake will now detect the host OS and use
appropriate NDK paths instead of assuming Linux.

- CMake_readelf.cmake will now be run with llvm-readelf from the NDK
toolchain when building for Android.

This has been tested on macOS and Linux. In principle the changes should
enable Android builds on Windows as well but I have not tested it so I
don't know if additional changes are required.

Issue: #2154