i#7442 skip csod: Add scheduler input mod by shard
The drmemtrace scheduler's input_thread_info_t modifiers use a 'tids'
field to allow applying the modifiers to a subset of a workload's
inputs.  But for core-sharded-on-disk traces there are no thread ids.
We solve that by adding a 'shards' field which can be used instead of
'tids' to specify inputs by ordinal (sorted lexicographically for a
directory's files, just like for 'only_shards').

Adds a unit test.

Issue: #7442, i#6685
Fixes #7442