i#7442 skip csod: Do not abort on a drmemtrace footer when skipping (#7443)
Since a core-shared-on-disk trace typically contains the footer record
from when a source software thread trace exited, the reader and its
skipping logic need to not abort when they see a footer. They don't need
to do anything on a footer as the EOF logic does not rely on such a
record. Here we fix both
reader_t::skip_instructions_with_timestamp() and
reader_t::process_input_entry() to not abort on a footer.

Adds a unit test, which fails without the fix.

Issue: #7442, #6685