i#7442 skip csod: Add scheduler input mod by shard (#7446)
The drmemtrace scheduler's input_thread_info_t modifiers use a 'tids'
field to allow applying the modifiers to a subset of a workload's
inputs. But for core-sharded-on-disk traces there are no thread ids. We
solve that by adding a 'shards' field which can be used instead of
'tids' to specify inputs by ordinal (sorted lexicographically for a
directory's files, just like for 'only_shards').

This change breaks binary compatibility, which unfortunately we cannot
maintain as C++'s offsetof is not reliable on this struct (due to some
implementations of std::set not being a standard-layout class). Adds a
release note.

Adds a unit test.

Issue: #7442, i#6685
Fixes #7442