i#7433:  properly initialize memref_t in test helper methods (#7435)
Fixes the problem of "initialized" memrefs containing uninitialized
fields.

In memref_gen.h, memref_t objects are initialized like this:

`memref_t memref = {}; 
`
However, since memref_t is a union, this form of aggregate
initialization only initializes the first declared union member. If the
first member doesn't cover all other union members, it may leave parts
of the other members uninitialized.

To fix this, a new first union member is added in the form of a byte
array sized to match the largest member struct. This ensures that `= {}`
style aggregate initialization of memref_t objects reliably
zero-initializes the entire union.

With this change in place, a seemingly unrelated unit test
(config_reader_unit_test.cpp) began failing a brittle equality test on
floating point values. Such FP comparisons are unsafe due to their
sensitivity to minor code perturbations, so a fix for this test is
included in this PR as well to enable the test suite to pass.

Fixes: #7433