i#6495 syscall inject: Set branch target marker at syscall trace end
Adds the TRACE_MARKER_TYPE_BRANCH_TARGET prior to the last instruction in the system call trace template. Also ensures that the last instruction is an indirect branch: eret on aarch64, sysret on x86, as observed on the traces collected on QEMU.

Sets the value of the TRACE_MARKER_TYPE_BRANCH_TARGET marker to the fallthrough pc of the syscall instruction, when injecting the trace template in raw2trace conversion.

Adds new invariant checks for the trace to ensure the last instr in the trace template is an indirect branch, with a value that is equal to the fallthrough pc of the pre-syscall-trace instruction (which is a syscall). There is an existing check that verifies PC continuity between the post-syscall-trace and pre-syscall-trace instruction.

Marks the aarch64 eret and x86 sysret as LINK_INDIRECT | LINK_RETURN and fixes behavior of instr_is_return and instr_is_mbr.

Issue: #6495