i#3683 android: Remove ARM-Android release package
Removes the release of an ARM-Android binary package since it does not
have good support for recent Android versions, partly because of the
lack of private library support.  The package can be re-added if that
support is later added; or it can be replaced with an AArch64 version
if that support becomes sufficient.  But right now it is misleading to
release this package as it implies it should work when most tools will
fail on recent Android versions.

Issue: #3683, #2154, #7312