i#6495 syscall inject: Handle x86 OP_sysret
Marks OP_sysret properly for instr_is_mbr, instr_is_return, and instr_branch_type.

sysret is not expected to be seen by the user-space DynamoRIO. We've seen it in kernel system call traces collected on QEMU as the instruction that ends the syscall trace and interrupts within the syscall trace. We want such instructions to have the TRACE_MARKER_TYPE_BRANCH_TARGET so they must be marked as an indirect branch.

Issue: #6495, #7157