i#6495 syscall inject: Handle AArch64 OP_eret (#7453)
Adds missing handling for OP_eret on AArch64. Adds the INSTR_CREATE_eret
macro, and marks OP_eret properly for instr_is_mbr, instr_is_return, and
instr_branch_type.

Adds unit test for eret IR decoding.

eret is not expected to be seen by the user-space DynamoRIO. We've seen
it in kernel system call traces collected on QEMU as the instruction
that ends the syscall trace and interrupts within the syscall trace. We
want such instructions to have the TRACE_MARKER_TYPE_BRANCH_TARGET so
they must be marked as an indirect branch.

Issue: #6495, #7157