i#6495 syscall inject: Handle x86 OP_sysret and OP_sysexit (#7454)
Marks the x86 OP_sysret and OP_sysexit properly for instr_is_mbr and
instr_branch_type.

sysret and sysexit are not expected to be seen by the user-space
DynamoRIO. We've seen sysret in kernel system call traces collected on
QEMU as the instruction that ends the syscall trace and interrupts
within the syscall trace. We want to annotate such instructions with the
TRACE_MARKER_TYPE_BRANCH_TARGET marker, which requires that they must be
marked as an indirect branch. This way we can show proper PC continuity
in the kernel syscall injected traces.

We do not mark sysret and sysexit as return instructions because they
are not returns in the usual sense that trace analysis tools would
expect. E.g., they don't read the return address from the stack, and
don't have matching call instrs (syscall and sysenter are not marked as
a call instr either, as they're not so from the point of view of the
user-mode DynamoRIO).

Issue: #6495, #7157