i#6495 syscall inject: Track syscall-fallthrough pc in raw2trace (#7458)
Today, raw2trace tracks the last pc if it was a syscall, which is used
to avoid consecutive duplicate system calls in the final trace.

Modifies raw2trace to instead track the fallthrough of the last pc if it
was a syscall. This works just as well for the duplicate syscall check,
and enables some upcoming logic which needs the fallthrough of the
syscall instruction.

The existing test_duplicate_syscalls in raw2trace_unit_tests still
passes.

Issue: #6495, #7157, #5934