Enable asserts in drmemtrace tests in release build
Many drmemtrace tests use assert() for their test logic.  Some undo
NDEBUG so they still perform checks in release build, but many do not.
We add an undef of NDEBUG for the onesmissing it, to match the others.
Maybe they should all use some locally defined ASSERT or CHECK but as
they are all already following an assert() convention this seems the
simplest path forward.