i#7157 dyn inject: Set syscall-end branch target marker
Sets the TRACE_MARKER_TYPE_BRANCH_TARGET marker if there is one present before an indirect branch instruction at the end the injected syscall trace.

Allows the invariant_checker to know whether dynamic injection of syscalls is enabled, so it can adjust the tests for ref/instr ordinals which are not incremented on the stream for dynamically injected records.

Adds an invocation of the invariant checker on a dynamically syscall-injected trace.

Checks the indirect_branch_target field of the syscall-end instr in an existing scheduler unit test.

Issue: #7157