i#6495 syscall inject: Set branch target marker in raw2trace (#7451)
Sets the `TRACE_MARKER_TYPE_BRANCH_TARGET` marker value if it is present
prior to the indirect branch at the last instruction position in the
system call trace template during raw2trace injection.

Sets the value of the `TRACE_MARKER_TYPE_BRANCH_TARGET` marker to the
fallthrough pc of the user-mode syscall instruction for which the trace
is being injected. It is harder to set it based on the actual next
instruction in the trace; the fallthrough pc works for all syscalls
except the control-transferring ones like sigreturn that we don't trace
anyway today (using PT or QEMU).

Adds new invariant checks for the injected trace: the last instruction
in the injected syscall trace should be an indirect branch (e.g., `eret`
on aarch64, `sysret` on x86, as observed on the syscall traces collected
on QEMU) with the target set to the next pc in the trace, only one
syscall trace should be injected after a user-space instr, and there
should not be any trace injected (for sigreturn) just prior to the
`TRACE_MARKER_TYPE_KERNEL_XFER` marker.

Adds unit tests for the new invariant checks, and modifies existing ones
to match the new expectations on syscall trace format (to have an
indirect branch as the syscall-trace-final instr). Also modifies the
burst_syscall_inject test to have the indirect branch at syscall trace
end in the generated template file. Also updates the
mock_syscall_sequences.x64 test file.

Documents the new requirements for the syscall template files in
trace_entry.h, and also allow `TRACE_MARKER_TYPE_BRANCH_TARGET` to have
public documentation as it is not really strictly internal anymore
(since users need to be aware of it and add it in the syscall template
file they generate).

Since the burst_syscall_inject test already runs the invariant_checker
on the raw2trace-injected trace, we know that the new invariants hold.

A later PR will add similar functionality to set the branch target
marker value during dynamic syscall trace injection in the drmemtrace
scheduler.

Issue: #6495