i#7464 msb memref: Allow non-memref to follow prefetch instr
Allows raw traces entries that do not look like OFFLINE_TYPE_MEMREF or OFFLINE_TYPE_MEMREF_HIGH to follow prefetch instructions during raw2trace conversion.

Prefetch instructions may generate a memref with an address that has some of the top 3 bits set which overwrites the type field to contain a value that is not either OFFLINE_TYPE_MEMREF or OFFLINE_TYPE_MEMREF_HIGH. Here, we override raw2trace logic to force it to process the raw trace entry following a prefetch instruction as a memref.

raw2trace conversion of a real trace failed with the "Unknown trace type" error without this fix. Reproduces the same issue on the builtin_prefetch app, which is now fixed.

Issue: #7464