Fix order of unreading split trace entries
Fixes the order in which split trace entries such as TRACE_MARKER_TYPE_SPLIT_VALUE are unread in unread_last_entry. TRACE_MARKER_TYPE_SPLIT_VALUE must be at the front of the pre_read so that it is returned _before_ the actual marker, in the same order it is read from the raw trace.