i#7470: Fix order of unreading split trace entries (#7469)
Fixes the order in which split trace entries such as
TRACE_MARKER_TYPE_SPLIT_VALUE are unread in unread_last_entry.
TRACE_MARKER_TYPE_SPLIT_VALUE must be at the front of the pre_read deque
so that it is returned _before_ the actual marker, in the same order it
is read from the raw trace.

Adds a raw2trace_unit_test that fails with "SPLIT_VALUE marker is not
adjacent to 2nd entry" due to the previously incorrect order of the
split_value marker wrt the actual marker in the deque. The error was
demonstrated using a large enough kernel_event marker value. This was
not seen on a real trace probably because kernel_event marker values do
not need the additional marker usually.

There is an additional bug where raw2trace does not fully take into
account the fact that some markers may be preceded by the
TRACE_MARKER_TYPE_SPLIT_VALUE marker. Added a TODO for now.

Issue: #7470