i#7157 dyn inject: Set syscall-end branch target marker (#7461)
Sets the value of the `TRACE_MARKER_TYPE_BRANCH_TARGET` marker to the
post-syscall pc, if there is one such marker present before an indirect
branch instruction at the end of an injected syscall trace.
    
Allows the invariant_checker to know whether dynamic injection of
syscalls is enabled, so it can adjust the tests for ref/instr ordinals
which are not incremented on the stream for dynamically injected
records.
    
Adds an invocation of the invariant checker on a dynamically
syscall-injected trace.
    
Checks the indirect_branch_target field of the syscall-end instr in an
existing scheduler unit test.
    
Issue: #7157