Mark schedule_stats_t::record_context_switch() virtual (#7479)
Makes the schedule_stats_t method record_context_switch() virtual so it
can be overridden in subclasses.

Addtiionally, since shard->prev_workload_id and shard->prev_tid are
cleared when this function is called,
we pass in the preserved values so there's no confusion.

Tested in an internal use case.