i#7303 drsyscall: Add a CI test to capture memory dump and syscalls. (#7481)
Add a CI test to capture memory dump mid-run and syscall records.

The test requires features from both runall and runmutli, so I move the
marco process_cmdline to a new file, and share it
between runall.cmake and runmulti.cmake. With this change, we can pass
precmd and postcmd for runall tests.

Issue: #7303