Enable asserts in drmemtrace tests in release build (#7460)
Many drmemtrace tests use assert() for their test logic. Some locally
undo NDEBUG so they still perform checks in release build, but many do
not. This PR adopts a more systematic approach: it disables the
re-addition of NDEBUG in the drcachesim CMake code for all tests, and
includes test_helpers.h in each test. To this header we add a sanity
check for NDEBUG to detect regressions.