i#3544 RV64: Replace usage of FindPythonInterp with FindPython3 in CMake
A Python 3 interpreter is required to generate instruction codecs for
RV64. Previously we make use of CMake's FindPythonInterp to search for
one, but the module has been deprecated in CMake since 3.12.

This patch replaces its usage with FindPython3 introduced in CMake 3.12,
which is fine since the minimum required CMake version of DynamoRIO has
been raised to 3.14. This fixes warnings about CMP0148 introduced since
CMake 3.27.

Issue: #3544