i#2297: AARCH64: Implement mbr & cbr instrumentation (#7005)
Implement mbr & cbr instrumentation for AARCH64, previously only
available for X86. For cbr instrumentation, there are three types of
conditional branches on AARCH64:

1. conditional branch if (non) zero: generate code to compare the
register against zero to compute the direction
2. test bit and branch if (non) zero: generate code to extract the bit
out of the register (using ubfm) and then compare the bit against zero
to compute the direction
3. conditional branch according to NZCV: generate code to compute the
direction using csinc instruction

Special care is implemented if the stolen register x28 is used as the
operand, to avoid generating incorrect results.

For mbr instrumentation, the target address is placed in the register,
which is later passed to the clean call from the user.

Additional instruction creation macros are added for ubfm and csinc
instructions. Support for immediate operand is added to the
INSTR_CREATE_eor macro. XINST_CREATE_slr_s is updated to use the new
INSTR_CREATE_ubfm macro.

Count-ctis test is now enabled due to cbr & mbr instrumentation
available on AARCH64. It has been enhanced to validate the direction and
target address of conditional branches to validate the instrumentation.

The instrcalls and cbrtrace example clients are enabled as well. They
are tested on real AARCH64 hardware. Additionally, a manually crafted
dynamorio client can successfully capture the whole control flow
transfer history with this pull reuqest.

Fixes: #2297 (only for AARCH64, ARM not yet) and #2919.