i#3544 RV64: Implement relink_special_ibl_xfer() (#7437)
Similar to AArch64, we cannot jump to an address storing in memory
directly on RISC-V, thus an instruction sequence of loading to scratch
(ld) and branching to register (jalr) must be used in special IBL. It's
only required to adjust the offset used in loading insturction during
relinking.

As drwrap-related tests aren't enabled on RV64, this commit is tested by
manually inspecting the relinked IBL in debugger and applications that
make use of shared libraries, for example, ld.bfd with LTO plugin, where
the address range is flushed on library unloading. Unit tests could be
introduced later.

Issue: #3544