i#7496: Allow syscall injection for sigreturn (#7497)
Adjusts injection point for the sigreturn syscall. For SYS_rt_sigreturn
we want to inject immediately before the TRACE_MARKER_TYPE_KERNEL_XFER
marker that is present just after the syscall func_arg markers (if any),
and before the post-syscall timestamp marker. Fixes this for both static
injection in raw2trace and dynamic injection in the drmemtrace
scheduler.

Removes the drmemtrace invariant check that disallowed trace injection
for sigreturn.

Adjusts the drmemtrace invariant checker to skip the
TRACE_MARKER_TYPE_BRANCH_TARGET marker check for sigreturns. This keeps
status-quo for signal return points where there is no marker that holds
the continuation point (note that the TRACE_MARKER_TYPE_KERNEL_XFER
marker also holds the post-sigreturn pc, not the resumption point); the
resumption point is already seen in the prior
TRACE_MARKER_TYPE_KERNEL_EVENT marker. Adds TODO (i#7496) to improve
this in future, but for now we skip on adding complexity that may not be
needed since the required information is already in the trace.

Relaxes an invariant check that disallowed func_arg markers after an
injected syscall trace without an intervening user-space instr. This may
happen in real traces when a signal arrives just before control enters a
function traced with `-record_function`, so the func_id-func_arg markers
for the traced function will be seen immediately after the signal
returns via a sigreturn (and the injected sigreturn trace).

Also adds more unit tests for raw2trace, scheduler, and the invariant
checker.

Verified on a large proprietary dynamically injected trace that there
are no invariant errors with injected SYS_rt_sigreturn traces.

Issue: #7496, #6495, #7157