i#7511 avx512: Identify ymm16+ as AVX-512 for lazy state preservation
DR's lazy AVX-512 state preservation was only looking for zmm and
opmask register usage: yet ymm16-31 are also AVX-512-only and my
machine has ymm16 code before any zmm code runs. DR fails to preserve
the state and the app state is corrupted.

The check is augmented here to also look for ymm16+.

The PREFIX_EVEX is documented to clarify it is deliberately only set
by decode_cti().

Tested:
My #7504 test's print() reliably printed the specifies ("%d")
instead of the values when under DR without the fix; with the fix it
prints the values every time.  Also confirmed in the logs with a client
with a bb event that the trigger happens where expected:
```
interp: start_pc = 0x00007fd93eb06980
  0x00007fd93eb06980  62 e2 7d 28 7a c6    vpbroadcastb {%k0} %esi -> %ymm16
Detected AVX-512 code in use
  0x00007fd93eb06986  89 f8                mov    %edi -> %eax
...
        zmm16= 0x0000000000d79a481c56f7360000000200000000000000003e9d71d800007fd73e9d75f800007fd73ea1878000007fd719ccdf1000007ffd19ccde5000007ffd
        zmm17= 0x0000000000d79a481c3d8f8a000000023e9d821800007fd7000002e0000000003e9d75f800007fd73ea1878000007fd719cce32800007ffd9d117c3b00005580
```

I considered adding a regression test by modifying avx512lazy.c to use
a ymm16 instruction: but that test is not useful on a machine where a
system library first uses ymm16.  We would need a pure-static test,
with a separate pure-static app per trigger.  I'm not sure it's worth
the effort and CI runtime to build and run all of those tests.

Fixes #7511