i#7504 avx512: Fix client.avx512lazy failures from system libs
As part of #7504 I investigated why the client.avx512lazy test fail on
my machine: it is related to #7504 in that it comes from AVX-512 code
in system libraries.  The test assumes the only such code will be
in our test application.  We look for that code and relax the assert in
that case. The test now passes when it failed every time on my machine.

Issue: #7504