i#7508: Add an option -memdump_on_window to the drmemtrace trace analysis tools framework. (#7509)
Add an option -memdump_on_window to the drmemtrace trace analysis tools
framework.

The option triggers a memory dump when a trace window opens by
-trace_after_instrs, -trace_instr_intervals_file, or
-retrace_every_instrs. If -retrace_every_instrs is also enabled, a
memory dump will be captured for each individual tracing window.

This feature only works on Linux X64. On Windows, raw2trace fails with
"Non-module instructions found with no encoding information." if it is
enabled. If this flag is used on Windows, the tracer will print an error
message without capturing a memory dump.

Add a new parameter, embed_timestamp, to get_unique_logfile() to embed a
timestamp in the file name. Adding a timestamp to the memory dump file
name allows us to better correlate the memory dump file to the
timestamps in the trace files.

Add a new CI test tool.drcacheoff.windows-memdump to test the new
option.

Issue: #7508