i#7516: Try in a loop if fd_priv_dup() fails
We observed internal applications failing in some environments with
-satisfy_w_xor_x with ENOMEM returned by the first two DUPFD attempts
in fd_priv_dup().  fd_priv_dup() only retries once and that value is
still to high in our failure cases.  Trying a 3rd time (/2 each time)
succeeds. Here we generalize for a loop that retries, to make things
more robust for cases where fds are limited.

Tested: Ran internal tests that were failing ~20% of the time and now
there are zero failures.

Fixes #7516