i#199: Add dr_invoke_syscall_as_app() (#7526)
Adds a new API routine dr_invoke_syscall_as_app() which invokes a system
call as though the application made it, meaning it goes through DR's
handling. However, it does not trigger client syscall events (skipping
them via an internal flag), which seems to be what use cases want (as
otherwise each client or ext has to use flags to skip them).

(We may want a separate dr_invoke_syscall() that is not treated as an
app syscall but just lets DR do its own internal tracking, but that is
harder to implement and not needed for my use case, so it is left as
future work for #199.)

Adds a test case that confirms DR acts on the syscall.

Issue: #199