i#7157 dyn inject: Add more tests for context switch injection (#7518)
Adds more unit tests for some context switch injection scenarios. This
PR is pure refactoring, with no functionality changes.

Adds testing to ensure that the stream APIs and the trace records are
consistent about the current tid, especially for the injected switch
trace.

Ensures that context switches happen only at the
`TRACE_MARKER_TYPE_CONTEXT_SWITCH_START` record when switch injection is
enabled.

Adds a record scheduler test that shows the current order of the new
input's tid-pid records inserted at context switches, wrt the injected
switch trace.

Also simplifies some logic in switch injection that caused some
confusion when reading the code.

Also fixes a test setup inconsistency for the context switch sequence,
which should not have any timestamp in the switch trace itself.

There is one pending TODO that will be addressed separately: moving the
synthetic tid-pid markers that are inserted at context switches to
before the injected switch trace. This is to match the behavior of the
stream APIs during dynamic injection, which identifies switch records by
the next input's tid (it's easier this way in online scheduling because
we already have switched to the new input when the switch records are
injected). Without this, the injected switch records are identified with
the prior input's tid and pid, which especially affects
core-sharded-on-disk traces.


Issue: #7157