Reset last_branch_target_ in the reader after its used
Resets the tracked last_branch_target_ in reader_t after we've created a memref using it. This is to avoid cases where the marker value is reused, which shouldn't ever happen. After this change, cases where there is a missing TRACE_MARKER_TYPE_BRANCH_TARGET before an indirect branch will clearly show a zero target instead of some non-zero value which is hard to judge sometimes.