Reset last branch target after usage in the reader (#7541)
Resets the tracked last_branch_target_ in reader_t after we've created a
memref using it. This is to avoid cases where the marker value is
reused, which shouldn't ever happen. After this change, cases where
there is a missing TRACE_MARKER_TYPE_BRANCH_TARGET before an indirect
branch will clearly show a zero target instead of some non-zero value,
which makes it easier to detect it as an error.