Set trace version in view_t with -skip_instrs
Ensures that the view_t sets the trace_version_ also when invoked with -skip_instrs which causes the TRACE_MARKER_TYPE_VERSION trace entry to not be seen by the tool. In this case we initialize it from the stream API instead; this is same as what it already does for file type.

Without this fix, the view tool does not print the target pc for indirect branches like ret.