i#7504 time scale: Add drx_register_time_scaling() (#7530)
Adds a new drx library feature: drx_register_time_scaling() which in
this initial implementation scales timers on Linux (both itimers and
POSIX timers) in both 64-bit and 32-bit.

Relaxes a DR assert to allow itimer processing during client init.
Clarifies the dr_syscall_set_param() docs: it does work to set state
during post-syscall events.

Adds a test of scaling itimers and POSIX timers in multiple scenarios
with respect to the timing of DR attach.
The test checks that the timer firing count is indeed scaled, and also
checks the post-syscall param is restored
and that queries receive the original value.

Issue: #7504