i#7504 time scale: Fix just-fired timer bug (#7545)
Fixes a bug where if a timer just fired right before
drx_unregister_time_scaling() queries it in order to restore it to its
app duration, the drx code will end up disabling the timer by leaving
the "it_value" as 0. This is fixed by resetting the timer ourselves by
setting "it_value" to "it_interval".

Removes client.drx_time_scale-test from the ignore list.

Tested:
The client.drx_time_scale-test was failing every 1 to 4 runs locally;
with this fix it now passes 50x in a row.
The test was also failing quite a bit on Github 32-bit which will serve
as a regression test.

Issue: #7504